load display and some starting on context menu
display the adding (which is added which is not added) problem is my version of the program do not have the IssueLoadScreen running after closing the AddComic.
I started context menu but EventHandler for menu item isn't working, menu item is not accepted in contextMenu.getItems().addAll. 
We need a remove function in localDB.
I implement the context menu in IssuePreview but that mean i can't update the tree in main.